changed terminology and renamed crate
"coerce" seems to be a better choice than "merge".
"Merging" gives the impression than we take two errors to make one,
while in fact we take one error of one of two types,
and lift (coerce) it to the most specific supertype.

Also fixed a few minor problems that I noticed
while reviewing the new wording.